Add base to support both export and link with projects
Now linking a project with JIRA must be done through a wizard instead
of the binding view. It must be either export (create the project on
jira) or link (link with an existing project on jira).

We keep this choice in the binding so we never export data to JIRA
when we chose the link only.