Skip import of tasks/worklogs before batch dates
Jira may send webhooks for old records, for instance worklogs from 6
month ago because their task has been changed. This change ensures
that we never import any record which has a last update date before the
last date of batch import.

The job methods now return the result of Importer.run() so the result is
shown on the jobs in the UI.