Partial revert of 2fa7d39
Commit title was: Skip import of tasks/worklogs before batch dates

The process is the following:
* Read T from backend, T is the is the last time we ran the import
* Create a batch job to import tasks/worklogs from T to now() and update
the backend with now()

The batch job is run asynchronously, it generates one job per task or
worklog to import between T and now(). When the jobs are executed, the
last batch date has already been updated to now(), so the jobs would
never import any record.