Refactorize aal batch importer with 'updating since' jira method
This Jira method returns the list of ids updated since a timestamp, and
the youngest timestamp, which means we can store this timestamp and
reuse it the next time we ask the modified worklogs: this way we should
never miss any update.

To support this, a type of field which support milliseconds in datetime
is required, otherwise lose this precision that Jira gives us.

As of today, the "updated since" method exists on for worklogs on the
Jira webservice, so the rest of the batch import still have to use the
JQL filters (with minute-precision only).